AAddress unused warnings
