+++ consensus/cryptarchia-sync/src/libp2p/mod.rs
@@ -1,6 +1,7 @@
 pub mod behaviour;
 mod downloader;
 pub mod errors;
+pub mod messages;
 mod packing;
 pub mod provider;
 mod utils;
